iotedge/tempmodule/tempSolution/modules/PythonModule/requirements.txt
azure-iot-device~=2.7.0
board
adafruit-circuitpython-dht
